requirements.txt
fastapi
uvicorn
python-dotenv
supabase
PyJWT
razorpay
requests